Moved BaseWrapper for SQL Object form og.meeting to og.base.
